+++ main.lua
@@ -1,21 +1,19 @@
--- Deluge v1.0.5
+-- Deluge v1.1.0
 -- Klehrik
 
 log.info("Successfully loaded ".._ENV["!guid"]..".")
 require("./helper")
 
 local enabled = false
-
-local director = nil
-local player = nil
-
-local stored_points = 0
 local stored_time = 0
-local point_increment = 0
+local stored_stages = 0
 local stored_health = 0
 local stored_max_health = 0     -- Make sure the health gained from max hp increase isn't affected by "healing" reduction
 local credits_init = false
 local current_char = 0
+
+local director = nil
+local player = nil
 
 local victory = {}
 for i = 1, 16 do table.insert(victory, 0) end
@@ -24,54 +22,63 @@ local char_names = {"Commando", "Huntress", "Enforcer", "Bandit", "HAN-D", "Engi
 local file_path = path.combine(paths.plugins_data(), "Klehrik-Deluge_sav2.txt")
 local succeeded, from_file = pcall(toml.decodeFromFile, file_path)
 if succeeded then
-    enabled = from_file.enabled
     victory = from_file.victory
 end
 
 local icon_path = _ENV["!plugins_mod_folder_path"].."/deluge.png"
-local diff_icon = gm.sprite_add(icon_path, 1, false, false, 25, 16)
-if diff_icon ~= -1 then log.info("Loaded difficulty icon sprite.") end
+local icon_path2x = _ENV["!plugins_mod_folder_path"].."/deluge2x.png"
+local diff_icon = gm.sprite_add(icon_path, 5, false, false, 12, 9)
+local diff_icon2x = gm.sprite_add(icon_path2x, 4, false, false, 25, 19)
+if diff_gray ~= -1 and diff_color ~= -1 then log.info("Loaded difficulty icon sprites.")
+else log.info("Failed to load difficulty icon sprites.") end
 
 
 -- Parameters
-local director_bonus = 0.5
+-- * Remember to update the description text below if modified
+local point_scaling = 0.5
 local speed_bonus = 0.25
 local healing_reduction = 0.5
+
+
+-- Initialize difficulty
+local diff_id = gm.difficulty_create("klehrik", "deluge")   -- namespace, identifier
+local class_diff = gm.variable_global_get("class_difficulty")[diff_id + 1]
+local values = {
+    "Deluge",       -- Name
+    "For those who have conquered Monsoon.\nYou will be washed away in a flood of pain.\n\n<c_stack>Director credits     + 50%\nEnemy move speed    + 25%\nAll healing            - 50%",  -- Description
+    diff_icon,      -- Sprite ID
+    diff_icon2x,    -- Sprite Loadout ID
+    7554098,        -- Primary Color
+    gm.constants.wUI_Select,    -- Sound ID
+    0.16,           -- diff_scale; Affects enemy stat scaling (health and damage)
+                    --     0.06 (Drizzle), 0.12 (Rainstorm), 0.16 (Monsoon)
+    3.0,            -- general_scale; Affects timer and chest price scaling
+                    --     The text update and bell sound only update/play at the start of every minute
+                    --     1.0 (Drizzle), 2.0 (Rainstorm), 3.0 (Monsoon)
+    1.7 * (1 + point_scaling),  -- point_scale; Affects point scaling, with the increase at any minute being 2.0 + (this * minutesElapsed)
+                    --                 e.g., with Monsoon's 1.7, at 5 minutes, the director gets 2.0 + (1.7 * 5) = 10.5 points per second
+                    --                 1.0 (Drizzle), 1.0 (Rainstorm), 1.7 (Monsoon)
+    true,           -- Either "is monsoon or higher" or "allow blight spawns"
+    true            -- Whichever one the bool above isn't
+}
+for i = 2, 12 do gm.array_set(class_diff, i, values[i - 1]) end
 
 
 
 -- ========== Main ==========
 
-local function dec_to_percent(value)
-    local str = tostring(gm.round(value * 100))
-    return gm.string_copy(str, 1, #str - 2)
-end
-
-
-local function save_to_file()
-    pcall(toml.encodeToFile, {
-        enabled = enabled,
-        victory = victory
-    }, {file = file_path, overwrite = true})
+local function does_instance_exist(inst)
+    return inst and gm._mod_instance_valid(inst) == 1.0
 end
 
 
 gui.add_imgui(function()
     if ImGui.Begin("Deluge") then
         -- Description
-        ImGui.Text("A fourth difficulty option, for those\nwho have already conquered Monsoon.\n\nDirector credits:           + "..dec_to_percent(director_bonus).."%\nEnemy move speed:    + "..dec_to_percent(speed_bonus).."%\nAll healing:                   - "..dec_to_percent(healing_reduction).."%\n\nYou must select Monsoon as a base\ndifficulty for Deluge modifiers to apply.\n\n(Can only be toggled while on\nthe character select screen)\n ")
-
-        -- Toggle status
-        local status = "DISABLED"
-        if enabled then status = "! ENABLED !" end
-        local can_toggle = find_cinstance_type(gm.constants.oSelectMenu)
-        if ImGui.Button("Status:  "..status) and can_toggle then
-            enabled = not enabled
-            save_to_file()
-        end
+        ImGui.Text("Adds a fourth difficulty option, for\nthose who have conquered Monsoon.")
 
         -- Victory display
-        ImGui.Text("\n_____________________________________\n\n[Deluge Victory Counts]\n ")
+        ImGui.Text("\n_____________________________________\n\n[Victory Counts]\n ")
         for i = 1, 15 do
             ImGui.Text(char_names[i]..":  "..victory[i])
         end
@@ -85,39 +92,38 @@ end)
 
 
 gm.pre_script_hook(gm.constants.__input_system_tick, function()
-    -- Reset when starting a new run (check for character select menu)
+
+    -- Reset some variables on the character select screen
     local select_ = find_cinstance_type(gm.constants.oSelectMenu)
     if select_ then
-        stored_points = 0
-        stored_time = 0
-        point_increment = 0
-        stored_health = 0
-        stored_max_health = 0
-        credits_init = false
+        enabled = false
         current_char = select_.choice
     end
 
 
-    if enabled then
-        if director and gm._mod_instance_valid(director) == 1.0 then
-            -- Run these every second
+    -- Check if Deluge is on
+    if gm._mod_game_getDifficulty() == diff_id then
+        enabled = true
+
+        if does_instance_exist(director) then
+
+            -- Reset variables when starting a new run
+            if director.time_start <= 0 then
+                stored_time = 0
+                stored_stages = 0
+                stored_health = 0
+                stored_max_health = 0
+                credits_init = false
+            end
+
+
+            -- Run this every second
             if stored_time < director.time_start then
-
-                -- Turn off if Monsoon is not actually on
-                if find_cinstance_type(gm.constants.oStageControl) and gm._mod_game_getDifficulty() ~= 2.0 then
-                    enabled = false
-                    save_to_file()
-                end
-
-
-                -- Get the director's point increment and increment again by a portion
                 stored_time = director.time_start
-                local inc = director.points - stored_points
-                if inc > 0 and stored_points > 0 then point_increment = inc * director_bonus end
 
-                stored_points = director.points + point_increment
-                director.points = director.points + point_increment
-
+                -- Increase points by another 1 point per second
+                -- This is because the 1.5x scaling from point_scale does not apply to the initial 2 pps
+                director.points = director.points + (2 * point_scaling)
 
                 -- Loop through all instances and look for enemies (on team 2.0)                
                 for i = 1, #gm.CInstance.instances_active do
@@ -138,10 +144,18 @@ gm.pre_script_hook(gm.constants.__input_system_tick, function()
 
 
         -- Remove 50% of all healing
-        if player and gm._mod_instance_valid(player) == 1.0 then
-            if stored_health > 0 and player.hp > stored_health and stored_max_health >= player.maxhp then
+        if does_instance_exist(player) then
+            -- Make sure the "healing" is not from stage transition
+            local stage_check = true
+            if does_instance_exist(director) and stored_stages < director.stages_passed then
+                stage_check = false
+                stored_stages = director.stages_passed
+            end
+
+            if stored_health > 0 and player.hp > stored_health and stored_max_health >= player.maxhp and stage_check then
                 player.hp = player.hp - ((player.hp - stored_health) * healing_reduction)
             end
+
             stored_health = player.hp
             stored_max_health = player.maxhp
 
@@ -165,26 +179,16 @@ gm.pre_script_hook(gm.constants.__input_system_tick, function()
             end
 
         end
+    end
 
 
-        -- Modify results screen a bit
+    -- Save victory
+    if enabled and find_cinstance_type(gm.constants.oCutscenePlayback2) and not credits_init then
         local results = find_cinstance_type(gm.constants.oResultsScreen)
         if results then
-
-            -- Victory or Defeat screen
-            -- (Doesn't actually work for the latter unfortunately)
-            if not results.is_history_screen then
-                results.diff_name = "Deluge"
-                results.diff_sprite = diff_icon
-            end
-
-            -- Save victory
-            if find_cinstance_type(gm.constants.oCredits) and (not credits_init) then
-                credits_init = true
-                victory[current_char + 1] = victory[current_char + 1] + 1
-                save_to_file()
-            end
-
+            credits_init = true
+            victory[current_char + 1] = victory[current_char + 1] + 1
+            pcall(toml.encodeToFile, {victory = victory}, {file = file_path, overwrite = true})
         end
     end
 end)